Refactored library code into several new modules. Updated most references to 'automated testing system' to 'Clout'. Renamed executable from 'run_test_suites.py' to 'clout'. Added tests/all_tests.py (based on PyNAST's all_tests.py). Added better usage example to clout executable and fixed up some of the docstrings. Changed copyright and version strings in all files.
