+++ clout/__init__.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+
+__author__ = "Jai Ram Rideout"
+__copyright__ = "Copyright 2012, The Clout Project"
+__credits__ = ["Jai Ram Rideout"]
+__license__ = "GPL"
+__version__ = "0.9-dev"
+__maintainer__ = "Jai Ram Rideout"
+__email__ = "jai.rideout@gmail.com"
+__status__ = "Development"
+
+__all__ = ['format', 'parse', 'run', 'util']
