+++ clout/parse.py
@@ -1,0 +1,96 @@
+#!/usr/bin/env python
+from __future__ import division
+
+__author__ = "Jai Ram Rideout"
+__copyright__ = "Copyright 2012, The Clout Project"
+__credits__ = ["Jai Ram Rideout"]
+__license__ = "GPL"
+__version__ = "0.9-dev"
+__maintainer__ = "Jai Ram Rideout"
+__email__ = "jai.rideout@gmail.com"
+__status__ = "Development"
+
+"""Module to parse various supported file formats."""
+
+def parse_config_file(config_f):
+    """Parses and validates a configuration file describing test suites.
+
+    Returns a list of lists containing the test suite label as the first
+    element and the command string needed to execute the test suite as the
+    second element.
+
+    Arguments:
+        config_f - the input configuration file describing test suites
+    """
+    results = []
+    used_test_suite_names = []
+    for line in config_f:
+        if not _can_ignore(line):
+            fields = line.strip().split('\t')
+            if len(fields) != 2:
+                raise ValueError("Each line in the config file must contain "
+                                 "exactly two fields separated by tabs.")
+            if fields[0] in used_test_suite_names:
+                raise ValueError("The test suite label '%s' has already been "
+                                 "used. Each test suite label must be unique."
+                                 % fields[0])
+            results.append(fields)
+            used_test_suite_names.append(fields[0])
+    if len(results) == 0:
+        raise ValueError("The config file must contain at least one test "
+                         "suite to run.")
+    return results
+
+def parse_email_list(email_list_f):
+    """Parses and validates a file containing email addresses.
+    
+    Returns a list of email addresses.
+
+    Arguments:
+        email_list_f - the input file containing email addresses
+    """
+    recipients = [line.strip() for line in email_list_f \
+                  if not _can_ignore(line)]
+    if len(recipients) == 0:
+        raise ValueError("There are no email addresses to send the test suite "
+                         "results to.")
+    for address in recipients:
+        if '@' not in address:
+            raise ValueError("The email address '%s' doesn't look like a "
+                             "valid email address." % address)
+    return recipients
+
+def parse_email_settings(email_settings_f):
+    """Parses and validates a file containing email SMTP settings.
+
+    Returns a dictionary with the key/value pairs 'smtp_server', 'smtp_port',
+    'sender', and 'password' defined.
+
+    Arguments:
+        email_settings_f - the input file containing tab-separated email
+            settings
+    """
+    required_fields = ['smtp_server', 'smtp_port', 'sender', 'password']
+    settings = {}
+    for line in email_settings_f:
+        if not _can_ignore(line):
+            try:
+                setting, val = line.strip().split('\t')
+            except:
+                raise ValueError("The line '%s' in the email settings file "
+                                 "must have exactly two fields separated by a "
+                                 "tab." % line)
+            if setting not in required_fields:
+                raise ValueError("Unrecognized setting '%s' in email settings "
+                                 "file. Valid settings are %r." % (setting,
+                                 required_fields))
+            settings[setting] = val
+    if len(settings) != 4:
+        raise ValueError("The email settings file does not contain one or "
+                "more of the following required fields: %r" % required_fields)
+    return settings
+
+def _can_ignore(line):
+    """Returns True if the line can be ignored (comment or blank line)."""
+    return False if line.strip() != '' and not line.strip().startswith('#') \
+           else True
