+++ templates/email_settings.txt
